# API/src/api.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Lockdown API
tags:
  - name: Get Environment
    description: Get all environments
  - name: World
    description: Only lockdown statuses for whole world
  - name: By Measures
    description: Specific NPIs for country
paths:
  /environment/{dsluid}/{module}:
    get:
      tags:
        - Environment
      summary: Environment
      path:
        - dsluid: ID of DSL
          description: ID of DSL
          in: int
          required: false
          schema:
            type: int
            format: int
      responses:
        "200":
          description: "List of Environemnts."
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    lockdown:
                      $ref: '#/components/schemas/Environment'
        default:
          description: Unexpected error
components:
  schemas:
    Measure:
      type: object
      properties:
        label:
          type: string
        name:
          type: string
        value:
          type: string
        '#date+start':
          type: string
          format: date
        '#date+end':
          type: string
          format: date
        '#meta+url':
          type: string
        '#country+code+iso3':
          type: string
    Environment:
      type: object
      properties:
        iso:
          type: string
        max_gathering: 
          type: array
          items:
            $ref: '#/components/schemas/Environment'
        measure:
          type: array
          items:
            $ref: '#/components/schemas/Environment'
        land:
          type: array
          items:
            $ref: '#/components/schemas/Environment'
        sea:
          type: array
          items:
            $ref: '#/components/schemas/Environment'
        flight:
          type: array
          items:
            $ref: '#/components/schemas/Environment'